Move Filters to RenderChain namespace
RenderChains require Filters to be of any use.
