+++ Examples/RenderScripts/Example.CustomTextures.cs
@@ -17,7 +17,6 @@
 using System;
 using System.Drawing;
 using Mpdn.Extensions.Framework;
-using Mpdn.Extensions.Framework.Filter;
 using Mpdn.Extensions.Framework.RenderChain;
 using Mpdn.RenderScript;
 using SharpDX;
