+++ Extensions/Framework/RenderChain/Filter.Sources.cs
@@ -15,9 +15,9 @@
 // License along with this library.
 
 using Mpdn.RenderScript;
-using IBaseFilter = Mpdn.Extensions.Framework.Filter.IFilter<Mpdn.IBaseTexture>;
+using IBaseFilter = Mpdn.Extensions.Framework.RenderChain.IFilter<Mpdn.IBaseTexture>;
 
-namespace Mpdn.Extensions.Framework.Filter
+namespace Mpdn.Extensions.Framework.RenderChain
 {
     public abstract class BaseSourceFilter<TTexture> : IFilter<TTexture>
         where TTexture : class, IBaseTexture
