+++ Extensions/Framework/RenderChain/RenderChain.cs
@@ -18,7 +18,6 @@
 using System;
 using System.Diagnostics;
 using System.IO;
-using Mpdn.Extensions.Framework.Filter;
 using Mpdn.OpenCl;
 using Mpdn.RenderScript;
 using TransformFunc = System.Func<System.Drawing.Size, System.Drawing.Size>;
