+++ Extensions/RenderScripts/Custom.MyRenderScript.cs
@@ -15,7 +15,6 @@
 // License along with this library.
 
 using System;
-using Mpdn.Extensions.Framework.Filter;
 using Mpdn.Extensions.Framework.RenderChain;
 using Mpdn.Extensions.RenderScripts.Mpdn.ImageProcessor;
 using Mpdn.Extensions.RenderScripts.Mpdn.Resizer;
