+++ Extensions/RenderScripts/Mpdn.Conditional.cs
@@ -16,7 +16,6 @@
 // 
 using System;
 using Mpdn.Extensions.Framework;
-using Mpdn.Extensions.Framework.Filter;
 using Mpdn.Extensions.Framework.RenderChain;
 using Mpdn.Extensions.Framework.Scripting;
