+++ Extensions/RenderScripts/Mpdn.Resizer.cs
@@ -17,7 +17,6 @@
 using System;
 using System.ComponentModel;
 using Mpdn.Extensions.Framework;
-using Mpdn.Extensions.Framework.Filter;
 using Mpdn.Extensions.Framework.RenderChain;
 using Mpdn.RenderScript;
